Time: 51 ms (13.76%) | Memory: 105.6 MB (8.99%) - LeetSync

+++ 2395-longest-binary-subsequence-less-than-or-equal-to-k/longest-binary-subsequence-less-than-or-equal-to-k.cpp
@@ -6,10 +6,10 @@ public:
         if(s[ind] == '0')
             return solve(ind-1, sz+1, ans, s, k);
         else{
-            if(pow(2, sz) + ans > k)
+            if(pow(2, sz)+ans > k)
                 return solve(ind-1, sz, ans, s, k);
             else
-                return solve(ind-1, sz+1, (int)pow(2, sz)+ans, s, k);
+                return solve(ind-1, sz+1, pow(2, sz)+ans, s, k);
         }
     }
     int longestSubsequence(string s, int k) {
